Introduce integration test modularity
