+++ openbb_platform/extensions/charting/integration/test_charting_python.py
@@ -1,6 +1,7 @@
 """Test charting extension."""
 
 import pytest
+from extensions.tests.conftest import parametrize
 from openbb_charting.core.openbb_figure import OpenBBFigure
 from openbb_core.app.model.obbject import OBBject
 
@@ -36,7 +37,7 @@ def get_equity_data():
     return data["stocks_data"]
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
@@ -59,7 +60,7 @@ def test_chart_equity_price_historical(params, obb):
     assert isinstance(result.chart.fig, OpenBBFigure)
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"symbol": "AAPL", "limit": 100, "chart": "True"}),
@@ -78,7 +79,7 @@ def test_chart_equity_fundamental_multiples(params, obb):
     assert isinstance(result.chart.fig, OpenBBFigure)
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
@@ -108,7 +109,7 @@ def test_chart_technical_adx(params, obb):
     assert isinstance(result.chart.fig, OpenBBFigure)
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
@@ -137,7 +138,7 @@ def test_chart_technical_aroon(params, obb):
     assert isinstance(result.chart.fig, OpenBBFigure)
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
@@ -167,7 +168,7 @@ def test_chart_technical_ema(params, obb):
     assert isinstance(result.chart.fig, OpenBBFigure)
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
@@ -197,7 +198,7 @@ def test_chart_technical_hma(params, obb):
     assert isinstance(result.chart.fig, OpenBBFigure)
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
@@ -228,7 +229,7 @@ def test_chart_technical_macd(params, obb):
     assert isinstance(result.chart.fig, OpenBBFigure)
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
@@ -259,7 +260,7 @@ def test_chart_technical_rsi(params, obb):
     assert isinstance(result.chart.fig, OpenBBFigure)
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
@@ -289,7 +290,7 @@ def test_chart_technical_sma(params, obb):
     assert isinstance(result.chart.fig, OpenBBFigure)
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
@@ -319,7 +320,7 @@ def test_chart_technical_wma(params, obb):
     assert isinstance(result.chart.fig, OpenBBFigure)
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
