+++ openbb_platform/extensions/derivatives/integration/test_derivatives_api.py
@@ -4,6 +4,7 @@ import base64
 
 import pytest
 import requests
+from extensions.tests.conftest import parametrize
 from openbb_core.env import Env
 from openbb_core.provider.utils.helpers import get_querystring
 
@@ -19,7 +20,7 @@ def headers():
     return {"Authorization": f"Basic {base64_bytes.decode('ascii')}"}
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"provider": "intrinio", "symbol": "AAPL", "date": "2023-01-25"}),
@@ -37,7 +38,7 @@ def test_derivatives_options_chains(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"symbol": "AAPL"}),
@@ -56,7 +57,7 @@ def test_derivatives_options_unusual(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
@@ -95,7 +96,7 @@ def test_derivatives_futures_historical(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"provider": "cboe", "symbol": "VX", "date": "2023-01-25"}),
