+++ openbb_platform/extensions/etf/integration/test_etf_api.py
@@ -2,6 +2,7 @@ import base64
 
 import pytest
 import requests
+from extensions.tests.conftest import parametrize
 from openbb_core.env import Env
 from openbb_core.provider.utils.helpers import get_querystring
 
@@ -18,7 +19,7 @@ def headers():
 # pylint: disable=redefined-outer-name
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"query": "", "provider": "fmp"}),
@@ -35,7 +36,7 @@ def test_etf_search(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
@@ -67,7 +68,7 @@ def test_etf_historical(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"symbol": "IOO", "provider": "fmp"}),
@@ -85,7 +86,7 @@ def test_etf_info(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"symbol": "IOO", "provider": "fmp"}),
@@ -103,7 +104,7 @@ def test_etf_sectors(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"symbol": "QQQ", "cik": None, "provider": "fmp"}),
@@ -120,7 +121,7 @@ def test_etf_holdings_date(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
@@ -152,7 +153,7 @@ def test_etf_holdings(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"symbol": "SPY,VOO,QQQ,IWM,IWN,GOVT,JNK", "provider": "fmp"})],
 )
@@ -167,7 +168,7 @@ def test_etf_price_performance(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"symbol": "IOO"})],
 )
@@ -182,7 +183,7 @@ def test_etf_countries(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"sort": "desc", "limit": 10})],
 )
@@ -197,7 +198,7 @@ def test_etf_discovery_gainers(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"sort": "desc", "limit": 10})],
 )
@@ -212,7 +213,7 @@ def test_etf_discovery_losers(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"sort": "desc", "limit": 10})],
 )
@@ -227,7 +228,7 @@ def test_etf_discovery_active(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"symbol": "qqq", "provider": "fmp"}),
