+++ openbb_platform/extensions/etf/integration/test_etf_python.py
@@ -1,5 +1,6 @@
 """Test etf extension."""
 import pytest
+from extensions.tests.conftest import parametrize
 from openbb_core.app.model.obbject import OBBject
 
 
@@ -16,7 +17,7 @@ def obb(pytestconfig):  # pylint: disable=inconsistent-return-statements
 # pylint: disable=redefined-outer-name
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"query": None, "provider": "fmp"}),
@@ -32,7 +33,7 @@ def test_etf_search(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
@@ -63,7 +64,7 @@ def test_etf_historical(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"symbol": "IOO", "provider": "fmp"}),
@@ -80,7 +81,7 @@ def test_etf_info(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"symbol": "IOO", "provider": "fmp"}),
@@ -97,7 +98,7 @@ def test_etf_sectors(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"symbol": "QQQ", "cik": None, "provider": "fmp"}),
@@ -113,7 +114,7 @@ def test_etf_holdings_date(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
@@ -144,7 +145,7 @@ def test_etf_holdings(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"symbol": "SPY,VOO,QQQ,IWM,IWN,GOVT,JNK", "provider": "fmp"})],
 )
@@ -158,7 +159,7 @@ def test_etf_price_performance(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"symbol": "IOO"})],
 )
@@ -172,7 +173,7 @@ def test_etf_countries(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"sort": "desc", "limit": 10})],
 )
@@ -186,7 +187,7 @@ def test_etf_discovery_gainers(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"sort": "desc", "limit": 10})],
 )
@@ -200,7 +201,7 @@ def test_etf_discovery_losers(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"sort": "desc", "limit": 10})],
 )
@@ -214,7 +215,7 @@ def test_etf_discovery_active(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"symbol": "qqq", "provider": "fmp"}),
@@ -231,7 +232,7 @@ def test_etf_holdings_performance(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"sort": "desc", "limit": 10})],
 )
@@ -245,7 +246,7 @@ def test_etf_discovery_gainers2(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"sort": "desc", "limit": 10})],
 )
@@ -259,7 +260,7 @@ def test_etf_discovery_losers2(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"sort": "desc", "limit": 10})],
 )
