+++ openbb_platform/extensions/fixedincome/integration/test_fixedincome_python.py
@@ -1,5 +1,6 @@
 """Test fixed income extension."""
 import pytest
+from extensions.tests.conftest import parametrize
 from openbb_core.app.model.obbject import OBBject
 
 # pylint: disable=redefined-outer-name
@@ -16,7 +17,7 @@ def obb(pytestconfig):
         return openbb.obb
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"start_date": "2023-01-01", "end_date": "2023-06-06"}),
@@ -30,7 +31,7 @@ def test_fixedincome_government_treasury_rates(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"date": "2023-01-01", "inflation_adjusted": True}),
@@ -44,7 +45,7 @@ def test_fixedincome_government_us_yield_curve(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"start_date": "2023-01-01", "end_date": "2023-06-06"}),
@@ -66,7 +67,7 @@ def test_fixedincome_sofr(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"start_date": "2023-01-01", "end_date": "2023-06-06"}),
@@ -88,7 +89,7 @@ def test_fixedincome_rate_estr(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"start_date": "2023-01-01", "end_date": "2023-06-06"}),
@@ -110,7 +111,7 @@ def test_fixedincome_rate_sonia(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"start_date": "2023-01-01", "end_date": "2023-06-06"}),
@@ -132,7 +133,7 @@ def test_fixedincome_rate_ameribor(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"start_date": "2023-01-01", "end_date": "2023-06-06"}),
@@ -154,7 +155,7 @@ def test_fixedincome_rate_effr(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({}),
@@ -169,7 +170,7 @@ def test_fixedincome_rate_effr_forecast(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"start_date": "2023-01-01", "end_date": "2023-06-06"}),
@@ -183,7 +184,7 @@ def test_fixedincome_rate_iorb(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"start_date": "2023-01-01", "end_date": "2023-06-06"}),
@@ -207,7 +208,7 @@ def test_fixedincome_rate_dpcredit(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
@@ -229,7 +230,7 @@ def test_fixedincome_rate_ecb(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"start_date": "2023-01-01", "end_date": "2023-06-06", "index_type": "yield"}),
@@ -257,7 +258,7 @@ def test_fixedincome_corporate_ice_bofa(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"start_date": "2023-01-01", "end_date": "2023-06-06", "index_type": "aaa"})],
 )
@@ -271,7 +272,7 @@ def test_fixedincome_corporate_moody(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
@@ -295,7 +296,7 @@ def test_fixedincome_corporate_commercial_paper(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
@@ -318,7 +319,7 @@ def test_fixedincome_corporate_spot_rates(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"date": "2023-01-01", "yield_curve": "spot"})],
 )
@@ -332,7 +333,7 @@ def test_fixedincome_corporate_hqm(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"start_date": "2023-01-01", "end_date": "2023-06-06", "maturity": "3m"})],
 )
@@ -346,7 +347,7 @@ def test_fixedincome_spreads_tmc(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"start_date": "2023-01-01", "end_date": "2023-06-06", "maturity": "10y"})],
 )
@@ -360,7 +361,7 @@ def test_fixedincome_spreads_tmc_effr(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"start_date": "2023-01-01", "end_date": "2023-06-06", "maturity": "3m"})],
 )
@@ -374,7 +375,7 @@ def test_fixedincome_spreads_treasury_effr(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"date": "2023-01-01", "yield_curve_type": "spot_rate"}),
@@ -398,7 +399,7 @@ def test_fixedincome_government_eu_yield_curve(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
