+++ openbb_platform/extensions/index/integration/test_index_api.py
@@ -2,6 +2,7 @@ import base64
 
 import pytest
 import requests
+from extensions.tests.conftest import parametrize
 from openbb_core.env import Env
 from openbb_core.provider.utils.helpers import get_querystring
 
@@ -18,7 +19,7 @@ def headers():
 # pylint: disable=redefined-outer-name
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"index": "dowjones", "provider": "fmp"})],
 )
@@ -33,7 +34,7 @@ def test_index_constituents(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
@@ -145,7 +146,7 @@ def test_index_market(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
@@ -187,7 +188,7 @@ def test_index_european(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"symbol": "BUKBUS", "provider": "cboe"})],
 )
@@ -202,7 +203,7 @@ def test_index_european_constituents(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"europe": True, "provider": "cboe"}),
@@ -221,7 +222,7 @@ def test_index_available(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"query": "D", "is_symbol": True, "provider": "cboe"}),
@@ -239,7 +240,7 @@ def test_index_search(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"provider": "cboe", "region": "US"})],
 )
@@ -254,7 +255,7 @@ def test_index_snapshots(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
