+++ openbb_platform/extensions/quantitative/integration/test_quantitative_api.py
@@ -5,6 +5,7 @@ from typing import Literal
 
 import pytest
 import requests
+from extensions.tests.conftest import parametrize
 from openbb_core.env import Env
 from openbb_core.provider.utils.helpers import get_querystring
 
@@ -64,7 +65,7 @@ def get_data(menu: Literal["equity", "crypto"]):
     return funcs[menu]()
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         ({"data": "", "target": "close"}, "equity"),
@@ -83,7 +84,7 @@ def test_quantitative_normality(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         ({"data": "", "target": "high"}, "equity"),
@@ -102,7 +103,7 @@ def test_quantitative_capm(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
@@ -137,7 +138,7 @@ def test_quantitative_omega_ratio(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         ({"data": "", "target": "close", "window": "5"}, "equity"),
@@ -156,7 +157,7 @@ def test_quantitative_kurtosis(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
@@ -191,7 +192,7 @@ def test_quantitative_unitroot_test(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         ({"data": "", "target": "close", "rfr": "", "window": ""}, "equity"),
@@ -210,7 +211,7 @@ def test_quantitative_sharpe_ratio(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
@@ -247,7 +248,7 @@ def test_quantitative_sortino_ratio(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         ({"data": "", "target": "close", "window": "220"}, "equity"),
@@ -265,7 +266,7 @@ def test_quantitative_skewness(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
@@ -300,7 +301,7 @@ def test_quantitative_quantile(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         ({"data": "", "target": "close"}, "equity"),
