+++ openbb_platform/extensions/regulators/integration/test_regulators_api.py
@@ -2,6 +2,7 @@ import base64
 
 import pytest
 import requests
+from extensions.tests.conftest import parametrize
 from openbb_core.env import Env
 from openbb_core.provider.utils.helpers import get_querystring
 
@@ -18,7 +19,7 @@ def headers():
 # pylint: disable=redefined-outer-name
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"symbol": "TSLA", "provider": "sec"}),
@@ -36,7 +37,7 @@ def test_regulators_sec_cik_map(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"query": "berkshire hathaway", "provider": "sec"}),
@@ -53,7 +54,7 @@ def test_regulators_sec_institutions_search(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"query": "2022", "provider": "sec", "url": ""}),
@@ -77,7 +78,7 @@ def test_regulators_sec_schema_files(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"query": "0000909832", "provider": "sec"}),
@@ -95,7 +96,7 @@ def test_regulators_sec_symbol_map(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"provider": "sec"})],
 )
@@ -110,7 +111,7 @@ def test_regulators_sec_rss_litigation(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"query": "oil", "use_cache": False, "provider": "sec"})],
 )
@@ -125,7 +126,7 @@ def test_regulators_sec_sic_search(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"query": "grain", "provider": "nasdaq"}),
@@ -142,7 +143,7 @@ def test_regulators_cftc_cot_search(params, headers):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
