+++ openbb_platform/extensions/regulators/integration/test_regulators_python.py
@@ -1,5 +1,6 @@
 """Test Regulators extension."""
 import pytest
+from extensions.tests.conftest import parametrize
 from openbb_core.app.model.obbject import OBBject
 
 
@@ -17,7 +18,7 @@ def obb(pytestconfig):
 # pylint: disable=redefined-outer-name
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"symbol": "TSLA", "provider": "sec"}),
@@ -33,7 +34,7 @@ def test_regulators_sec_cik_map(params, obb):
     assert isinstance(result.results.cik, str)
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"query": "berkshire hathaway", "provider": "sec"}),
@@ -47,7 +48,7 @@ def test_regulators_sec_institutions_search(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"query": "2022", "provider": "sec", "url": None}),
@@ -68,7 +69,7 @@ def test_regulators_sec_schema_files(params, obb):
     assert len(result.results.files) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"query": "0000909832", "provider": "sec"}),
@@ -84,7 +85,7 @@ def test_regulators_sec_symbol_map(params, obb):
     assert isinstance(result.results.symbol, str)
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"provider": "sec"})],
 )
@@ -96,7 +97,7 @@ def test_regulators_sec_rss_litigation(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [({"query": "oil", "use_cache": False, "provider": "sec"})],
 )
@@ -108,7 +109,7 @@ def test_regulators_sec_sic_search(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         ({"query": "grain", "provider": "nasdaq"}),
@@ -122,7 +123,7 @@ def test_regulators_cftc_cot_search(params, obb):
     assert len(result.results) > 0
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params",
     [
         (
