+++ openbb_platform/extensions/technical/integration/test_technical_api.py
@@ -5,6 +5,7 @@ from typing import Literal
 
 import pytest
 import requests
+from extensions.tests.conftest import parametrize
 from openbb_core.env import Env
 from openbb_core.provider.utils.helpers import get_querystring
 
@@ -66,7 +67,7 @@ def get_crypto_data():
     return data["crypto_data"]
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
@@ -106,7 +107,7 @@ def test_technical_atr(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
@@ -146,7 +147,7 @@ def test_technical_fib(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         ({"data": "", "index": "", "offset": ""}, "equity"),
@@ -166,7 +167,7 @@ def test_technical_obv(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         ({"data": "", "index": "", "length": "", "signal": ""}, "equity"),
@@ -186,7 +187,7 @@ def test_technical_fisher(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
@@ -224,7 +225,7 @@ def test_technical_adosc(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
@@ -266,7 +267,7 @@ def test_technical_bbands(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
@@ -304,7 +305,7 @@ def test_technical_zlma(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         ({"data": "", "index": "", "length": "", "scalar": ""}, "equity"),
@@ -332,7 +333,7 @@ def test_technical_aroon(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
@@ -370,7 +371,7 @@ def test_technical_sma(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
@@ -410,7 +411,7 @@ def test_technical_demark(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         ({"data": "", "index": "", "anchor": "", "offset": ""}, "equity"),
@@ -430,7 +431,7 @@ def test_technical_vwap(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
@@ -470,7 +471,7 @@ def test_technical_macd(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
@@ -508,7 +509,7 @@ def test_technical_hma(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
@@ -546,7 +547,7 @@ def test_technical_donchian(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
@@ -588,7 +589,7 @@ def test_technical_ichimoku(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         ({"data": "", "index": "date", "target": "close", "period": "10"}, "equity"),
@@ -616,7 +617,7 @@ def test_technical_clenow(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         ({"data": "", "index": "", "offset": ""}, "equity"),
@@ -636,7 +637,7 @@ def test_technical_ad(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
@@ -674,7 +675,7 @@ def test_technical_adx(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
@@ -712,7 +713,7 @@ def test_technical_wma(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         ({"data": "", "index": "", "length": "", "scalar": ""}, "equity"),
@@ -740,7 +741,7 @@ def test_technical_cci(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
@@ -780,7 +781,7 @@ def test_technical_rsi(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
@@ -818,7 +819,7 @@ def test_technical_stoch(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
@@ -858,7 +859,7 @@ def test_technical_kc(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         ({"data": "", "index": "", "length": ""}, "equity"),
@@ -878,7 +879,7 @@ def test_technical_cg(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
@@ -920,7 +921,7 @@ def test_technical_cones(params, data_type):
     assert result.status_code == 200
 
 
-@pytest.mark.parametrize(
+@parametrize(
     "params, data_type",
     [
         (
