+++ openbb_platform/extensions/tests/test_docstring_examples.py
@@ -1,6 +1,7 @@
 """Test utils."""
 import pytest
-from integration.utils.helpers import check_docstring_examples
+
+from extensions.tests.utils.helpers import check_docstring_examples
 
 
 @pytest.mark.integration
